Refactor day 6
Game logic:
Added const and noexcpect where needed

Separated enemy bullet and player bullet, which resulted in removing an indentation in collision logic

removed all comments

reduced the length for the high score functions. Made it more performant. It used an insert therefore it was o(n) perfomance now we use pushback which is o(1)

reduced the length of the mega function that checkscollision, took help from chatgpt

reduced the line length of DeleteDeadEntities()

Removed the usage of char* and insted use std::string or std::string_view